requirement.txt
# Core dependencies for AI News Processor
# Recommended Python version: 3.9-3.11

# Web interface
gradio>=4.40.0,<5.0.0

# OpenAI API
openai>=1.40.0,<2.0.0

# AI Agent framework
git+https://github.com/openai/swarm.git

# Web scraping and search
requests>=2.28.0,<3.0.0
beautifulsoup4>=4.11.0,<5.0.0
googlesearch-python>=1.2.0,<2.0.0

# Environment management
python-dotenv>=1.0.0,<2.0.0

# HTML parsing (additional support)
lxml>=4.9.0,<5.0.0

# HTTP client improvements
urllib3>=1.26.0,<3.0.0

# Optional: Enhanced error handling
tenacity>=8.2.0,<9.0.0